+++ test/repository/ComponentRepositoryTests.java
@@ -5,15 +5,17 @@ import static helpers.builders.DonationBuilder.aDonation;
 import static helpers.builders.DonorBuilder.aDonor;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+
 import java.util.Arrays;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
+
 import model.component.Component;
 import model.component.ComponentStatus;
 import model.donation.Donation;
 import model.donor.Donor;
+
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+
 import suites.ContextDependentTestSuite;
 
 public class ComponentRepositoryTests extends ContextDependentTestSuite {
